ACCUMULO-3828 Repeatedly check constraint in proxy tests
Also make the test cleanup method a bit smarter.

+++ test/src/test/java/org/apache/accumulo/proxy/SimpleProxyBase.java
@@ -201,7 +201,9 @@ public abstract class SimpleProxyBase extends SharedMiniClusterIT {
   public void teardown() throws Exception {
     if (null != table) {
       try {
-        client.deleteTable(creds, table);
+        if (client.tableExists(creds, table)) {
+          client.deleteTable(creds, table);
+        }
       } catch (Exception e) {
         log.warn("Failed to delete test table", e);
       }
@@ -1113,6 +1115,12 @@ public abstract class SimpleProxyBase extends SharedMiniClusterIT {
     writerOptions.setThreads(1);
     writerOptions.setTimeoutMs(100000);
 
+    Map<String,Integer> constraints = client.listConstraints(creds, table);
+    while (!constraints.containsKey(NumericValueConstraint.class.getName())) {
+      log.info("Constraints don't contain NumericValueConstraint");
+      Thread.sleep(2000);
+    }
+
     String batchWriter = client.createWriter(creds, table, writerOptions);
     client.update(batchWriter, mutation("row1", "cf", "cq", "x"));
     client.update(batchWriter, mutation("row1", "cf", "cq", "x"));
@@ -1126,6 +1134,12 @@ public abstract class SimpleProxyBase extends SharedMiniClusterIT {
     } catch (MutationsRejectedException e) {}
 
     client.removeConstraint(creds, table, 2);
+
+    constraints = client.listConstraints(creds, table);
+    while (constraints.containsKey(NumericValueConstraint.class.getName())) {
+      log.info("Constraints still contains NumericValueConstraint");
+      Thread.sleep(2000);
+    }
 
     assertScan(new String[][] {}, table);
